Browser support: Indicate iOS 7 is no longer tested
We had to drop iOS 7 testing as iOS simulators are no longer supported
by BrowserStack and there are no real devices for iOS <10. We continue
to rely on old iOS emulators for testing as long as they work. iOS 8.3 & 9.3
seem to work fine for now but iOS 7.0 was breaking a lot; right now, the
browser doesn't even start.

Closed gh-229