Index: restore "Resources" block style, remove JQF/Other projects
The "Resources" section felt a bit lost on the home page due to the
large padding and lack of block styling on its heading.
It used to have a custom styling via `id="sidebar"` which was similar
to a block, but was lost at some point without replacement.
https://web.archive.org/web/20130301123913/jquery.com

Change it to a proper `<h2 class="block">` to look more on par with
the rest of the page, and also remove the `padding-left: 60px` from
the .resources class which is redundant now.

Also:
* remove outdated "About the jQuery Foundation" link.

* remove "Other Related Projects" section, which used low-quality
  images, was broken on mobile, loaded an extra copy of these logos
  on the home page, gave high-prominence to deprecated jQuery Mobile,
  and maintenance-only jQuery UI, used an outdated version of the
  QUnit logo.

  Perhaps the global navigation which points to the same projects
  already, suffices?

  By removing these, we bring the first code example above the fold,
  which seems valuable. If we keep it, I'd suggest moving it to the
  bottom of the content section at least.